fixed based on the comments
